Testing adding and playing mp3 files

+++ tests/dbtest.sh
@@ -2,5 +2,9 @@
 
 ${PREFIX} ${bindir:-.}/dbtest ${srcdir:-.}/testdata/file1.ogg "file one"
 result=$?
+test $result -eq 0 || exit $result
 rm empty.db
-exit $result
+${PREFIX} ${bindir:-.}/dbtest ${srcdir:-.}/testdata/file2.mp3 "file two"
+result=$?
+test $result -eq 0 || exit $result
+rm empty.db
